move EVENT from constants.ts to common.ts
