+++ src/api/websocket/common.ts
@@ -1,1 +1,13 @@
 export type WebSocketCallback<T> = (data: T) => void;
+
+export const EVENT = {
+  // /users
+  USER_STATUS_SET: 'user-status:set',
+  USER_STATUS_RECEIVE: 'user-status:receive',
+  // /channels
+  MESSAGE_SEND: 'message:send',
+  MESSAGE_RECEIVE: 'message:receive',
+  MESSAGE_RECEIVE_ALL: 'message:receive-all',
+  CHANNEL_JOIN: 'channel:join',
+  CHANNEL_LEAVE: 'channel:leave',
+};
