+++ src/api/websocket/useChannel.ts
@@ -1,6 +1,6 @@
 import { useCallback, useContext } from 'react';
-import { EVENT } from '../../components/config/constants';
 import { SocketContext } from '../../contexts/SocketContext';
+import { EVENT } from './common';
 
 const useChannel = () => {
   const { client } = useContext(SocketContext);
