+++ src/api/websocket/useMessage.ts
@@ -1,9 +1,9 @@
 import { useCallback, useContext } from 'react';
 import { UserStatusEnum } from '../generated';
-import { EVENT } from '../../components/config/constants';
+import { EVENT, WebSocketCallback } from './common';
 import { SocketContext } from '../../contexts/SocketContext';
-import { WebSocketCallback } from './common';
 
+// TODO けす
 type ReceiveMessagePayload = {
   status: UserStatusEnum;
   userID: number;
